Add function to calculate distance

+++ models/dashboard.go
@@ -3,6 +3,6 @@ package models
 type Dashboard struct {
 	FullName        string  `json:"full_name"`
 	RoleDescription string  `json:"role_description"`
-	OfficeLongitude float32 `json:"office_longitude"`
-	OfficeLatitude  float32 `json:"office_latitude"`
+	OfficeLongitude float64 `json:"office_longitude"`
+	OfficeLatitude  float64 `json:"office_latitude"`
 }
